Move non block tests to separate source file
