+++ tests/ring/CMakeLists.txt
@@ -2,6 +2,7 @@ project(ring_test LANGUAGES C)
 
 add_executable(${PROJECT_NAME}
         ring_test.c
+        non_blocking_ring_suite.c
         ../../src/non_blocking_ring.c
 )
